restructure to pipeline framework
