+++ projects/streaming+diffpy/diffstreamz/diffstreamz/gaussianfitlib.py
@@ -22,7 +22,8 @@ their starting values.
 
 from __future__ import print_function
 from diffpy.srfit.fitbase import FitContribution, FitRecipe, Profile, FitResults
-
+import numpy as np
+import matplotlib.pyplot as plt
 
 class GaussianFit(object):
     '''Least-squares fit of Gauss function to the specified data.
@@ -181,3 +182,43 @@ class GaussianFit(object):
         return
 
 # end of class GaussianFit
+
+# Define functions that are used in gaussian fitting workflows
+
+plt.ion()
+fig = plt.figure()
+ax1 = fig.add_subplot(211)
+ax2 = fig.add_subplot(212)
+ 
+def generate_data(x0):
+    x = np.arange(-20, 20, 0.1)
+    sig = 1.5
+    noise = 0.2 * np.ones_like(x)
+    y = np.exp(-0.5*(x-x0)**2/sig**2) + noise * np.random.randn(*x.shape)
+    plt.pause(1)
+    dat = (x, y)
+    return dat
+
+def plot_data(dat):
+    x, y = dat[0], dat[1]
+    #f1 = plt.figure(1)
+    #plt.ion(); plt.clf()
+    ax1.cla()
+    line1 = ax1.plot(x, y, 'x')
+    fig.canvas.draw()
+    plt.pause(0.1)
+
+def fit_data(dat):
+    x, y = dat[0], dat[1]
+    gfit = GaussianFit(x,y)
+    gfit.refine()
+    plt.pause(1)
+    return gfit
+
+def plot_fit(gfit):
+    ax2.cla()
+    l2a = ax2.plot(gfit.x, gfit.y, 'b.', label="observed Gaussian")
+    l2b = ax2.plot(gfit.x, gfit.yg, 'g-', label="calculated Gaussian")
+    ax2.legend()
+    fig.canvas.draw()
+    plt.pause(0.1)
